Create Jumping off point...in Python
Oye.  I know we discussed Perl, I tried it and failed miserably, I looked into ImageMagick and Perl..I cant figure it out, I tried for a bit and it was like pulling shards of glass out of my face. I'm gonna give it another go; but so far in python its pretty easy.  I'm gonna try to recreate this in Perl.  

So what its doing is pretty simple so far, using the PIL (Python image library) library.  Opens an image, gets the size, gets a pixel value at (1,2), compares the pixel value to a test to see if it matches and prints match.