Update 11.22 NEW AND IMPROVED.
Added some comments and added two more 'databases' of skin tones.
Will need some more testing of course; but this is the meat and potatoes of our project.  We just need to polish it up and add things like directory querying for search (maybe crawling of computer), flagging of file above X% skin pixels.  Shouldnt be too hard but this is enough to submit as final tool; work will be done during the break.